+++ packages/cli/plugin-tailwind/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @modern-js/plugin-tailwindcss
 
+<<<<<<< HEAD
 ## 1.20.1
 
 ### Patch Changes
@@ -21,6 +22,8 @@
   - @modern-js/utils@1.20.0
   - @modern-js/runtime@1.20.0
 
+=======
+>>>>>>> 6ab74256f (Release v1.19.0 (#1791))
 ## 1.19.0
 
 ### Patch Changes